Much prettier
I yoinked some of my font loading and drawing code out of my nms-glyph project to pretty up the output a little.
I also added some more information to the image.
The embedded font seems to be public domain, or CC0, or whatever? Included the license to the best of my abilities.